secure

+++ resources/views/admin/accessories/edit.blade.php
@@ -33,8 +33,11 @@
                 <div class="form-group row mt-4">
                     <label for="inputDesc" class="col-sm-2 control-label">الوصف </label>
                     <div class="col-sm-10">
-                        <textarea class="form-control @error('description') is-invalid @enderror" id="inputDesc" rows="3" placeholder="أدخل وصف المنتج" name='description'>{{ old('description',$accessory->description) }}</textarea>
-                        @error('description')<div class="invalid-feedback">{{ $message }}</div>@enderror
+                        <textarea class="form-control @error('description') is-invalid @enderror" id="inputDesc"
+                                  rows="3" placeholder="أدخل وصف المنتج"
+                                  name='description'>{{ old('description',$accessory->description) }}</textarea>
+                        @error('description')
+                        <div class="invalid-feedback">{{ $message }}</div>@enderror
                     </div>
                 </div>
 
@@ -42,22 +45,27 @@
                 <div class="form-group row mt-4">
                     <label for="inputCategory" class="col-sm-2 control-label">القسم </label>
                     <div class="col-sm-10">
-                        <select class="form-control select2 @error('category_id') is-invalid @enderror" id="inputCategory" name="category_id" style="width: 100%">
+                        <select class="form-control select2 @error('category_id') is-invalid @enderror"
+                                id="inputCategory" name="category_id" style="width: 100%">
                             @foreach($categories as $category)
-                                <option value="{{ $category->id }}" {{$accessory->category_id == $category->id ?'selected':''}}>
+                                <option
+                                    value="{{ $category->id }}" {{$accessory->category_id == $category->id ?'selected':''}}>
                                     {{ $category->name }}
                                 </option>
                             @endforeach
                         </select>
-                        @error('category_id')<div class="invalid-feedback">{{ $message }}</div>@enderror
+                        @error('category_id')
+                        <div class="invalid-feedback">{{ $message }}</div>@enderror
                     </div>
                 </div>
-               <!-- سعر المنتج -->
+                <!-- سعر المنتج -->
                 <div class="form-group row mt-4">
-                    <label for="price" class="col-sm-2 control-label">سعر المنتج  </label>
+                    <label for="price" class="col-sm-2 control-label">سعر المنتج </label>
                     <div class="col-sm-10">
-                        <input class="form-control @error('price') is-invalid @enderror" id="price" name="price" style="width: 100%" value="{{$accessory->price}}"/>
-                        @error('price')<div class="invalid-feedback">{{ $message }}</div>@enderror
+                        <input class="form-control @error('price') is-invalid @enderror" id="price" name="price"
+                               style="width: 100%" value="{{$accessory->price}}"/>
+                        @error('price')
+                        <div class="invalid-feedback">{{ $message }}</div>@enderror
                     </div>
                 </div>
                 <!-- نوع الخصم    -->
@@ -66,9 +74,17 @@
                     <div class="col-sm-10">
                         <select name="discount_type" id="discount_type"
                                 class="form-control select2 @error('discount_type') is-invalid @enderror">
-                            <option value="" {{ old('discount_type', $accessory->discount?->discount_type) == '' ? 'selected' : '' }}>لا يوجد</option>
-                            <option value="fixed" {{ old('discount_type', $accessory->discount?->discount_type) == 'fixed' ? 'selected' : '' }}>ثابت</option>
-                            <option value="percentage" {{ old('discount_type', $accessory->discount?->discount_type) == 'percentage' ? 'selected' : '' }}>نسبة
+                            <option
+                                value="" {{ old('discount_type', $accessory->discount?->discount_type) == '' ? 'selected' : '' }}>
+                                لا يوجد
+                            </option>
+                            <option
+                                value="fixed" {{ old('discount_type', $accessory->discount?->discount_type) == 'fixed' ? 'selected' : '' }}>
+                                ثابت
+                            </option>
+                            <option
+                                value="percentage" {{ old('discount_type', $accessory->discount?->discount_type) == 'percentage' ? 'selected' : '' }}>
+                                نسبة
                                 مئوية
                             </option>
                         </select>
@@ -82,8 +98,9 @@
                     <div class="col-sm-10">
                         <input type="number" step="0.01"
                                class="form-control @error('discount_value') is-invalid @enderror" id="discount_value"
-                               placeholder="أدخل قيمة الخصم" name='discount_value'  value="{{ old('discount_value', $accessory?->discount?->discount_value) }}""
-                               min="0">
+                               placeholder="أدخل قيمة الخصم" name='discount_value'
+                               value="{{ old('discount_value', $accessory?->discount?->discount_value) }}""
+                        min="0">
                         @error('discount_value')
                         <div class="invalid-feedback">{{ $message }}</div>@enderror
                     </div>
@@ -104,7 +121,8 @@
                     <label for="end_date" class="col-sm-2 control-label">تاريخ نهاية الخصم </label>
                     <div class="col-sm-10">
                         <input type="date" class="form-control @error('end_date') is-invalid @enderror" id="end_date"
-                               placeholder="أدخل تاريخ نهاية الخصم" name='end_date' value="{{ old('end_date', $accessory?->discount?->end_date?->format('Y-m-d')) }}">
+                               placeholder="أدخل تاريخ نهاية الخصم" name='end_date'
+                               value="{{ old('end_date', $accessory?->discount?->end_date?->format('Y-m-d')) }}">
                         @error('end_date')
                         <div class="invalid-feedback">{{ $message }}</div>@enderror
                     </div>
@@ -113,8 +131,11 @@
                 <div class="form-group row">
                     <label for="quantity" class="col-sm-2 control-label"> الكمية المتاحة</label>
                     <div class="col-sm-10">
-                        <input type="number" class="form-control @error('quantity') is-invalid @enderror" id="quantity" placeholder="أدخل  كمية المنتج" name='quantity' value="{{$accessory->quantity}}" required min="1" step="1">
-                        @error('quantity')<div class="invalid-feedback">{{ $message }}</div>@enderror
+                        <input type="number" class="form-control @error('quantity') is-invalid @enderror" id="quantity"
+                               placeholder="أدخل  كمية المنتج" name='quantity' value="{{$accessory->quantity}}" required
+                               min="1" step="1">
+                        @error('quantity')
+                        <div class="invalid-feedback">{{ $message }}</div>@enderror
                     </div>
                 </div>
 
@@ -133,16 +154,19 @@
                     <div class="col-sm-10">
                         <div id="imagePreviewContainer" class="d-flex flex-wrap">
                             <div class="m-2" id="img-preview">
-                                @foreach($accessory->images as $key => $image)
-                                    <div style="display: inline-block; position: relative;">
-                                        <img src="{{ asset('storage/'.$image) }}" class="img-thumbnail"
-                                             style="height: 100px; width: 100px; object-fit: cover;">
-{{--                                        <button type="button" onclick="deleteImage({{ $key }})"--}}
-{{--                                                style="position: absolute; top: 5px; right: 5px; background-color: red; color: white; border: none;">--}}
-{{--                                            x--}}
-{{--                                        </button>--}}
-                                    </div>
-                                @endforeach
+                                @if($accessory->images)
+                                    @foreach ($accessory->images as $key => $image)
+                                        <div style="display: inline-block; position: relative;">
+                                            <img src="{{ asset('storage/'.$image) }}" class="img-thumbnail"
+                                                 style="height: 100px; width: 100px; object-fit: cover;">
+                                            {{--                                        <button type="button" onclick="deleteImage({{ $key }})"--}}
+                                            {{--                                                style="position: absolute; top: 5px; right: 5px; background-color: red; color: white; border: none;">--}}
+                                            {{--                                            x--}}
+                                            {{--                                        </button>--}}
+                                        </div>
+
+                                    @endforeach
+                                @endif
                             </div>
                             <div id="newImagePreviewContainer" class="d-flex flex-wrap mt-2">
                                 <!-- معاينات الصور الجديدة -->
@@ -169,13 +193,13 @@
             previewContainer.html(''); // مسح المعاينات السابقة
 
             if (event.target.files) {
-                $.each(event.target.files, function(_, file) {
+                $.each(event.target.files, function (_, file) {
                     var reader = new FileReader();
-                    reader.onload = function(e) {
+                    reader.onload = function (e) {
                         var imgWrapper = $('<div>').addClass('m-2');
                         var img = $('<img>').attr('src', e.target.result)
                             .addClass('img-thumbnail')
-                            .css({ height: '100px', width: '100px', objectFit: 'cover' });
+                            .css({height: '100px', width: '100px', objectFit: 'cover'});
                         imgWrapper.append(img);
                         previewContainer.append(imgWrapper);
                     }
